requirements.txt
Flask>=2.3
gunicorn
python-dotenv
google-genai 
